requirements.txt
Django==3.2.7
djangorestframework==3.12.4
django-cors-headers==3.10.1
gunicorn==20.1.0
pytz==2021.1
Pillow==8.3.2